Move some functions from PBclust to PBlib
Functions that could be used out of the context of the PBclust.py
command line were moved to PBlib. The functions that appeared specific
to the command line (mostly I/O functions) remain in PBclust.
